Move lex_routes into LexPipe in lex.ts
