+++ lib/syntax/lex.js
@@ -89,3 +89,19 @@ var LexSpace = (function (_super) {
 }(Lex));
 exports.LexSpace = LexSpace;
 ;
+var lex_routes = {
+    " ": new LexSpace(),
+    "#": new LexComment(),
+    "“": new LexString(),
+};
+var LexPipe = (function (_super) {
+    __extends(LexPipe, _super);
+    function LexPipe(out) {
+        var _this = this;
+        lex_routes[frames_1.Frame.kOUT] = out;
+        _this = _super.call(this, lex_routes) || this;
+        return _this;
+    }
+    return LexPipe;
+}(frames_1.Frame));
+exports.LexPipe = LexPipe;
