+++ lib/syntax/pipeline.d.ts
@@ -5,8 +5,5 @@ export declare class EvalPipe extends Frame {
 export declare class ParsePipe extends Frame {
     constructor(out: Frame, meta?: Context);
 }
-export declare class LexPipe extends Frame {
-    constructor(out: Frame, meta?: Context);
-}
 export declare const framify: (input: string, context?: Context) => Frame;
 export declare const framify_new: (input: string, context?: Context) => Frame;
