+++ lib/syntax/pipeline.js
@@ -6,11 +6,6 @@ var __extends = (this && this.__extends) || function (d, b) {
 };
 var frames_1 = require("../frames");
 var lex_1 = require("./lex");
-var lex_routes = {
-    " ": new lex_1.LexSpace(),
-    "#": new lex_1.LexComment(),
-    "“": new lex_1.LexString(),
-};
 var EvalPipe = (function (_super) {
     __extends(EvalPipe, _super);
     function EvalPipe(out, meta) {
@@ -35,25 +30,13 @@ var ParsePipe = (function (_super) {
     return ParsePipe;
 }(frames_1.Frame));
 exports.ParsePipe = ParsePipe;
-var LexPipe = (function (_super) {
-    __extends(LexPipe, _super);
-    function LexPipe(out, meta) {
-        if (meta === void 0) { meta = frames_1.Void; }
-        var _this = this;
-        meta[frames_1.Frame.kOUT] = out;
-        _this = _super.call(this, meta) || this;
-        return _this;
-    }
-    return LexPipe;
-}(frames_1.Frame));
-exports.LexPipe = LexPipe;
 var piper = function (input, context) {
     if (context === void 0) { context = frames_1.Void; }
     var source = new frames_1.FrameString(input);
     var result = new frames_1.FrameArray([], context);
     var evaluator = new EvalPipe(result);
     var parser = new ParsePipe(evaluator);
-    var lexer = new LexPipe(parser, lex_routes);
+    var lexer = new lex_1.LexPipe(parser);
     var status = source.reduce(lexer);
     if (status !== lexer) {
         console.error("\n* pipe returned " + status);
@@ -79,7 +62,7 @@ exports.framify_new = function (input, context) {
 var pipe = function (input, out) {
     var source = new frames_1.FrameString(input);
     var output = new frames_1.FrameArray([]);
-    var router = new LexPipe(output, lex_routes);
+    var router = new lex_1.LexPipe(output);
     var status = source.reduce(router);
     if (status !== router) {
         console.error("\n* pipe returned " + status);
