+++ src/syntax/pipeline.ts
@@ -1,12 +1,6 @@
 import * as _ from "lodash";
 import { Context, Frame, FrameArray, FrameLazy, FrameString, FrameSymbol, Void } from "../frames";
-import { Lex, LexComment, LexSpace, LexString } from "./lex";
-
-const lex_routes = {
-  " ": new LexSpace(),
-  "#": new LexComment(),
-  "“": new LexString(),
-};
+import { LexPipe } from "./lex";
 
 export class EvalPipe extends Frame {
   constructor(out: Frame, meta: Context = Void) {
@@ -22,19 +16,12 @@ export class ParsePipe extends Frame {
   }
 }
 
-export class LexPipe extends Frame {
-  constructor(out: Frame, meta: Context = Void) {
-    meta[Frame.kOUT] = out;
-    super(meta);
-  }
-}
-
 const piper = (input: string, context = Void): Frame => {
   const source = new FrameString(input);
   const result = new FrameArray([], context); // store the result
   const evaluator = new EvalPipe(result); // evaluate expressions in context
   const parser = new ParsePipe(evaluator); // assemble tokens into expressions
-  const lexer = new LexPipe(parser, lex_routes); // convert string into tokens
+  const lexer = new LexPipe(parser); // convert string into tokens
 
   const status = source.reduce(lexer);
   if (status !== lexer) {
@@ -62,7 +49,7 @@ export const framify_new = (input: string, context = Void): Frame => {
 const pipe = (input: string, out: Frame): Frame => {
   const source = new FrameString(input);
   const output = new FrameArray([]);
-  const router = new LexPipe(output, lex_routes);
+  const router = new LexPipe(output);
 
   const status = source.reduce(router);
   if (status !== router) {
